minor fixes for front page
