+++ packages/website/src/features/Header/Header.tsx
@@ -74,7 +74,13 @@ export const Header = () => {
   const breakpoint = useBreakpoint();
 
   return (
-    <Box bg="black" borderBottom="1px solid" borderColor="gray.700">
+    // height is statically set here to prevent cls
+    <Box
+      bg="black"
+      borderBottom="1px solid"
+      borderColor="gray.700"
+      height="56px"
+    >
       <Flex align="center" pt={[4, 4, 0]} px={3} flexWrap="wrap">
         <Link
           href={links.HOMEPAGE}
